requirements.txt
asgiref==3.8.1
astroid==3.0.3
black==23.12.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
Django==5.1.4
filelock==3.16.1
flake8==7.0.0
identify==2.6.3
isort==5.13.2
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.6.0
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.0.3
python-dotenv==1.0.1
PyYAML==6.0.2
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.28.0
